Remove population of duid field
This `duid` field was once a "multi-platform" field on ingress into mParticle, and would be translated to a different platform specific property on egress. This has long since changed, and this field now only represents the "android_id" field of the Events API. Roku needs to stop populating this field.